requirements.txt
Package                 Version
----------------------- -----------
hydra-core              1.1.1
matplotlib              3.5.1
pandas                  1.4.1
plotly                  5.6.0
pytorch-lightning       1.6.3
torch                   1.10.2
torchaudio              0.10.2
torchmetrics            0.7.2
torchvision             0.11.3
wandb                   0.12.10
